requirements.txt
numpy>=1.20.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
matplotlib>=3.3.0
torch>=1.8.0
typing-extensions>=3.7.4
PyYAML>=5.4.1
tqdm>=4.65.0